requirements.txt
rich
google-api-python-client
google-auth-oauthlib
